Fix MCP server tests and CI configuration
- Move MCP server tests that call external chart API to integration tests
- Update CI configuration to remove moved tests from smoke tests
- MCP integration tests will run in dedicated CI job with chart API service
- Add database cleanup to queue service tests to prevent test pollution
- Fix queue service test query to work without locking for testing

Progress on CI failures:
✅ MCP server tests properly moved to integration tests
✅ CI configuration updated to run MCP tests with chart API service
🔄 Queue service tests still failing - investigating query issues
