+++ .github/workflows/ci-core.yml
@@ -141,7 +141,7 @@ jobs:
     - name: Run fast unit tests (no database)
       run: |
         export PATH="$HOME/.cargo/bin:$PATH"
-        cargo test --lib -- --test-threads=4 --nocapture config auth::services metrics mcp_server::tests::test_mcp_server_creation mcp_server::tests::test_get_data_sources mcp_server::tests::test_get_series_catalog mcp_server::tests::test_create_fallback_visualization mcp_server::tests::test_call_private_chart_api_success mcp_server::tests::test_call_private_chart_api_failure mcp_server::tests::test_get_series_data_for_visualization mcp_server::tests::test_search_economic_series_with_custom_limit mcp_server::tests::test_search_economic_series_missing_query mcp_server::tests::test_get_series_data_missing_series_id mcp_server::tests::test_get_series_data_with_custom_limit mcp_server::tests::test_create_data_visualization_missing_series_ids mcp_server::tests::test_create_data_visualization_with_default_chart_type mcp_server::tests::test_create_data_visualization_with_title mcp_server::tests::test_create_data_visualization_without_title mcp_server::tests::test_create_data_visualization_with_date_filters
+        cargo test --lib -- --test-threads=4 --nocapture config auth::services metrics mcp_server::tests::test_mcp_server_creation mcp_server::tests::test_get_data_sources mcp_server::tests::test_get_series_catalog mcp_server::tests::test_create_fallback_visualization mcp_server::tests::test_get_series_data_for_visualization mcp_server::tests::test_search_economic_series_with_custom_limit mcp_server::tests::test_search_economic_series_missing_query mcp_server::tests::test_get_series_data_missing_series_id mcp_server::tests::test_get_series_data_with_custom_limit mcp_server::tests::test_create_data_visualization_missing_series_ids mcp_server::tests::test_create_data_visualization_with_default_chart_type mcp_server::tests::test_create_data_visualization_with_title mcp_server::tests::test_create_data_visualization_without_title mcp_server::tests::test_create_data_visualization_with_date_filters
       working-directory: backend
       env:
         RUST_LOG: warn
